optimalStrategy 동전이 2개만 있는경우 까지 구해봄

+++ DataStructureWithAlgorithm/src/main/java/Algorithm/DynamicProgramming/블럭채우기/CodeUp3709.java
@@ -4,21 +4,21 @@ import java.util.Scanner;
 
 public class CodeUp3709 {
 
-    static int[] dp = new int[10001];
+    static long[] dp = new long[10001];
     static int DIV = 100000007;
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
+        long n = sc.nextLong();
         CodeUp3709 s = new CodeUp3709();
-        int r = s.solution(n);
+        long r = s.solution(n);
         System.out.println(r);
     }
 
-    private int solution(int n) {
+    private long solution(long n) {
         if(n == 1) return 1;
         else if(n == 2) return 2;
-        else if(dp[n] != 0) return dp[n];
-        return dp[n] = (solution(n - 1) + solution(n - 2)) % DIV;
+        else if(dp[(int) n] != 0) return dp[(int) n];
+        return dp[(int) n] = (solution(n - 1) + solution(n - 2)) % DIV;
     }
 }
